8- Sistema_Biblioteca/BotLivros/requirements.txt
altgraph==0.17.4
arrow==1.3.0
asttokens==2.4.1
attrs==24.2.0
auto-py-to-exe==2.44.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.8.2
botcity-email-plugin==0.1.5
botcity-framework-base==0.4.4
botcity-framework-core==1.2.1
botcity-framework-web==0.9.0
botcity-http-plugin==0.1.4
botcity-maestro-sdk==0.5.1
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
contourpy==1.2.1
cookiecutter==2.6.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
Eel==0.17.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
fonttools==4.53.1
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.10
imap-tools==1.7.2
importlib_metadata==8.5.0
ipykernel==6.29.4
ipython==8.26.0
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msedge-selenium-tools==3.141.3
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.48
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pynput==1.7.7
pyparsing==3.1.2
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-xlib==0.33
pytz==2024.2
pywin32==306
pywin32-ctypes==0.2.2
pywinauto==0.6.8
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
reportlab==4.2.4
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
selenium==3.141.0
setuptools==70.3.0
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
text-unidecode==1.3
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
wcwidth==0.2.13
webdriver-manager==4.0.2
Werkzeug==3.0.4
whichcraft==0.6.1
widgetsnbextension==4.0.11
zipp==3.20.2
zope.event==5.0
zope.interface==6.4.post2
